requirements.txt
nose2
sphinx
scrapy
statsd
mock